1628 rings

+++ Noise.h
@@ -89,7 +89,7 @@ void SetupBlackAndBlueStripedPalette()
 
 void drawNoise(CRGBPalette16 palette, uint8_t hueReduce = 0)
 {
-  for (uint8_t i = 0; i < NUM_LEDS; i++) {
+  for (uint16_t i = 0; i < NUM_LEDS; i++) {
     uint8_t x = coordsX[i];
     uint8_t y = coordsY[i];
